refactor: pass all interfaces to declare events for hubtype analytics in botonic/core
