+++ packages/botonic-core/src/models/index.ts
@@ -1,3 +1,4 @@
 export * from './events'
+export * from './hubtype-analytics'
 export * from './legacy-types'
 export * from './user'
